refactor: extract camera hooks and components
